Use the directory that was created
Great project, thank you for this!

One thing I noticed: https://earthstar-project.org/tutorials/run-a-server contains instructions to "Create a new directory and name it `.share_data` (starting with a dot)."
But then the code uses `"./data"` for what seems to be storing share data, right? Was this by mistake? If so, here's a PR! :)